+++ packages/backend-skia/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @glissade/backend-skia
 
+## 0.58.1-pre.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @glissade/scene@0.58.1-pre.0
+  - @glissade/core@0.58.1-pre.0
+
 ## 0.58.0
 
 ### Patch Changes