+++ packages/vite-plugin/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @glissade/vite-plugin
 
+## 0.58.1-pre.0
+
+### Patch Changes
+
+- @glissade/core@0.58.1-pre.0
+
 ## 0.58.0
 
 ### Patch Changes